requirements.txt
Django==5.2.5
gunicorn==23.0.0
boto3==1.40.21
django-storages==1.14.6
whitenoise==6.9.0
python-dotenv==1.1.1